+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 * Add autosave feature for `blueberry4$` users
 * Add `unpause_eggs` feature
 * Turn "Toggle Powerups" into "Enable Powerups"
+* Add `keep_powerups_enabled` settings, because they seem to revert back to their original state
 
 ## 0.1.3 (2018-03-17)