+++ CHANGELOG.md
@@ -1,11 +1,12 @@
-## 0.1.4 (WIP)
+## 0.1.4 (2018-03-26)
 
 * Add autosave feature for `blueberry4$` users
 * Add `unpause_eggs` feature
 * Turn "Toggle Powerups" into "Enable Powerups"
-* Add `keep_powerups_enabled` settings, because they seem to revert back to their original state
+* Add `keep_powerups_enabled` settings, because they revert back to their original state after a manual save
 * Add action to line up all eggs by the incubator
-* Add action & setting to backup creatures
+* Add action & setting for backing up creatures
+* Many fixes from the `creatures` library
 
 ## 0.1.3 (2018-03-17)